+++ src/main/java/com/likelion/wisesaying/function/saying/End.java
@@ -1,6 +1,6 @@
-package com.likelion.wisesaying.function;
+package com.likelion.wisesaying.function.saying;
 
-import com.likelion.wisesaying.controller.IMainControllable;
+import com.likelion.wisesaying.function.IMainControllable;
 import com.likelion.wisesaying.repository.obj.SayingRepository;
 import com.likelion.wisesaying.service.SayingService;
 import java.util.Map;
