+++ src/main/java/com/likelion/wisesaying/function/saying/Update.java
@@ -1,9 +1,8 @@
-package com.likelion.wisesaying.function;
+package com.likelion.wisesaying.function.saying;
 
-import com.likelion.wisesaying.controller.IMainControllable;
 import com.likelion.wisesaying.domain.Saying;
+import com.likelion.wisesaying.function.IMainControllable;
 import com.likelion.wisesaying.language.KoreaContent;
-import com.likelion.wisesaying.util.exception.CustomRequestException;
 import com.likelion.wisesaying.util.request.Request;
 import java.util.Map;
 
@@ -18,19 +17,17 @@ public class Update implements IMainControllable {
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             return;
-        } catch (CustomRequestException e) {
-            return;
         }
 
         System.out.println(KoreaContent.UPDATE_REQUEST_CONTENT+ saying.getContent());
         System.out.print(KoreaContent.REQUEST_CONTENT);
         String updateContent = Request.input();
-        saying.setContent(updateContent);
 
         System.out.println(KoreaContent.UPDATE_REQUEST_AUTHOR + saying.getAuthor());
         System.out.print(KoreaContent.REQUEST_AUTHOR);
         String updateAuthor = Request.input();
-        saying.setAuthor(updateAuthor);
+
+        saying.updateSaying(updateContent, updateAuthor);
 
         service.update(saying);
     }
